organize model.utils
split this out into separate files
