+++ pyPhenology/utils.py
@@ -108,7 +108,7 @@ def load_saved_model(filename):
     if not isinstance(filename, str):
         raise TypeError('filename must be string, got' + type(filename))
     
-    model_info = models.utils.read_saved_model(filename)
+    model_info = models.utils.misc.read_saved_model(filename)
         
     if model_info['model_name']=='BootstrapModel':
         # The bootstrap model has it's own code for loading saved files
